pyproject.toml
[project]
name = "minimax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "eventlet>=0.40.4",
    "flask>=3.1.3",
    "flask-socketio>=5.6.1",
    "python-chess>=1.999",
    "stockfish>=4.0.8",
]
